Non-spatial clustering

A function containing two steps used at different times in the clustSIGNAL workflow. An initial non-spatial clustering and sub-clustering step (reclust = FALSE) is used to generate groups of ‘putative cell types’, whereas a later non-spatial clustering step (reclust = TRUE) is used to cluster adaptively smoothed gene expression data.

Usage

nsClustering(spe, dimRed, reclust, ...)

Arguments

spe

SpatialExperiment object. For reclust = FALSE, the object should contain logcounts and PCA, but for reculst = TRUE, the object should contain smoothed gene expression.

reclust

a logical parameter handled within the method.

Value

SpatialExperiment object containing 'putative cell type' group allotted to each cell (reclust = FALSE) or clusters generated from smoothed data (reclust = TRUE).

Examples

data(mouseEmbryo2)

spe <- scater::runPCA(spe)

# reclust = FALSE for non-spatial clustering of normalised counts
# reclust = TRUE for non-spatial clustering of adaptively smoothed counts
spe <- nsClustering(spe, dimRed = "PCA", reclust = FALSE)
#> [1] "Initial nonspatial clustering performed. Clusters = 17 2024-08-27 01:24:29.320558"
#> [1] "Nonspatial subclustering performed. Subclusters = 82 2024-08-27 01:24:35.039882"
head(spe$nsCluster)
#> [1] 12 12 12 12 7  12
#> Levels: 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17
head(spe$nsSubcluster)
#>  embryo2_Pos29_cell10_z2  embryo2_Pos29_cell10_z5 embryo2_Pos29_cell100_z2 
#>                     12.3                     12.3                     12.3 
#> embryo2_Pos29_cell100_z5 embryo2_Pos29_cell101_z2 embryo2_Pos29_cell101_z5 
#>                     12.3                      7.1                     12.3 
#> 82 Levels: 1.1 1.2 1.3 1.4 1.5 1.6 1.7 1.8 10.1 10.2 10.3 11.1 11.2 ... 9.8