clustSIGNAL

A clustering method for cell type classification of spatial transcriptomics data. The tool generates and uses an adaptively smoothed, spatially informed gene expression data for clustering.

Usage

clustSIGNAL(
  spe,
  samples,
  cells,
  dimRed = "None",
  NN = 30,
  kernel = "G",
  spread = 0.05,
  sort = TRUE,
  threads = 1,
  outputs = "c",
  ...
)

Arguments

spe

a SpatialExperiment object.

samples

a character indicating name of colData(spe) column containing sample names.

cells

a character indicating name of colData(spe) column containing cell IDs.

dimRed

a character indicating the name of the reduced dimensions to use from the SpatialExperiment object (i.e., from reducedDimNames(spe)). Default value is 'PCA'.

NN

an integer for the number of neighbourhood cells the function should consider. The value must be greater than or equal to 1. Default value is 30.

kernel

a character for type of distribution to be used. The two valid values are "G" or "E". G for Gaussian distribution, and E for exponential distribution. Default value is "G".

spread

a numeric value for distribution spread, represented by standard deviation for Gaussian distribution and rate for exponential distribution. Default value is 0.05 for Gaussian distribution and 20 for exponential distribution.

sort

a logical parameter for whether or not to sort the neighbourhood after region description. Default value is TRUE.

threads

a numeric value for the number of CPU cores to be used for the analysis. Default value set to 4 cores.

outputs

a character for the type of output to return to the user. "c" for data frame of cell IDs and their respective cluster numbers (default), "n" for list of dataframe of clusters plus neighbourhood matrix, "s" for list of dataframe of clusters plus final spatialExperiment object, or "a" for list of all outputs.

...

additional parameters for TwoStepParam clustering methods. Include parameters like k for number of nearest neighbours and cluster.fun for selecting community detection method. Default values k = 5, cluster.fun = "louvain".

Value

a list of outputs

clusters: a data frame of cell names and their cluster classification.

neighbours: a matrix of cell names and the names of their NN nearest neighbour cells.

spe_final: a SpatialExperiment object with initial 'putative cell type' groups, entropy values, smoothed gene expression, post-smoothing clusters, and silhouette widths included.

Examples

data(mouseEmbryo2)
#> Warning: data set ‘mouseEmbryo2’ not found

names(colData(spe))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'colData': object 'spe' not found
# identify the column names with cell and sample labels
samples = "sample_id"
cells = "uniqueID"
res_list <- clustSIGNAL(spe, samples, cells, outputs = "c")
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'spatialCoords': object 'spe' not found