Domainness measure

A function to measure the heterogeneity of a cell's neighbourhood in terms of entropy, such that homogeneous neighbourhoods have low entropy and heterogeneous neighbourhoods have high entropy.

Usage

entropyMeasure(spe, cells, regXclust, threads)

Arguments

spe

SpatialExperiment object with logcounts, PCA, and 'putative cell type' groups included.

cells

a character vector of cell IDs of each cell. Length of vector must be equal to the number of cells in spatialExperiment object (i.e. the number of rows in colData(spe)).

regXclust

a list of vectors for each cell's neighbourhood composition indicated by the proportion of 'putative cell type' groups it contains.

threads

a numeric value for the number of CPU cores to be used for the analysis.

Value

SpatialExperiment object including entropy values for each cell neighbourhood.

Examples

data(mouseEmbryo2)

# requires list of containing cluster proportions of each region (regXclust),
# generated using the neighbourDetect() function
spe <- entropyMeasure(spe, cells = "uniqueID", regXclust, threads = 1)
#> Error in eval(expr, envir, enclos): object 'regXclust' not found
head(spe$entropy)
#> NULL