Non-spatial clustering

A function containing two steps used at different times in the clustSIGNAL workflow. An initial non-spatial clustering and sub-clustering step (reclust = FALSE) is used to generate groups of ‘putative cell types’, whereas a later non-spatial clustering step (reclust = TRUE) is used to cluster adaptively smoothed gene expression data.

Usage

nsClustering(spe, samples, dimRed, batch, reclust, ...)

Arguments

spe

SpatialExperiment object. For reclust = FALSE, the object should contain logcounts and PCA, but for reculst = TRUE, the object should contain smoothed gene expression.

samples

a character indicating name of colData(spe) column containing sample names.

dimRed

a character indicating the name of the reduced dimensions to use from the SpatialExperiment object (i.e., from reducedDimNames(spe)). Default value is 'PCA'.

batch

a logical parameter for whether or not to perform batch correction. Default value is FALSE.

reclust

a logical parameter handled within the method.

...

additional parameters for TwoStepParam clustering methods. Include parameters like k for number of nearest neighbours and cluster.fun for selecting community detection method. Default values k = 5, cluster.fun = "louvain".

Value

SpatialExperiment object containing 'putative cell type' group allotted to each cell (reclust = FALSE) or clusters generated from smoothed data (reclust = TRUE).

Examples

data(example)

# reclust = FALSE for non-spatial clustering of normalised counts
# reclust = TRUE for non-spatial clustering of adaptively smoothed counts
spe <- nsClustering(spe, samples = "sample_id", dimRed = "PCA", batch = FALSE,
                    reclust = FALSE)
#> [1] "Initial nonspatial clustering performed. Clusters = 7 Time 00:22:18"
#> Warning: You're computing too large a percentage of total singular values, use a standard svd instead.
#> [1] "Nonspatial subclustering performed. Subclusters = 19 Time 00:22:19"
head(spe$nsCluster)
#> [1] 5 5 6 5 5 5
#> Levels: 1 2 3 4 5 6 7
head(spe$nsSubcluster)
#> embryo2_Pos29_cell110_z2 embryo2_Pos29_cell117_z2 embryo2_Pos29_cell128_z2 
#>                      5.2                      5.2                      6.3 
#> embryo2_Pos29_cell134_z2  embryo2_Pos29_cell14_z2 embryo2_Pos29_cell141_z2 
#>                      5.2                      5.2                      5.2 
#> 19 Levels: 1.1 1.2 2.1 2.2 2.3 3.1 3.2 3.3 4.1 4.2 4.3 5.1 5.2 5.3 6.1 ... 7.2